+++ src/main/java/org/ednovo/gooru/client/mvp/play/collection/end/CollectionEndView.java
@@ -1338,7 +1338,7 @@ public class CollectionEndView extends BaseViewWithHandlers<CollectionEndUiHandl
 
 
 	@Override
-	public void displaySpendTime(Long hours, Long mins, Long secs) {
+	public void displaySpendTime(Long hours, Long mins, Double secs) {
 		spendTimeContainer.clear();
 		if(AppClientFactory.isAnonymous()){
 			dispalyTime();
